requirements.txt
alembic==1.7.5
aniso8601==9.0.1
backports.entry-points-selectable==1.1.0
click==7.1.2
distlib==0.3.3
filelock==3.3.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
platformdirs==2.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyGreSQL==5.2.3
PyMySQL==1.0.2
pytz==2021.3
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.26
virtualenv==20.8.1
webargs==8.0.1
Werkzeug==2.0.2
PyJWT==2.3.0
python-dotenv==0.20.0
bcrypt==3.2.0
onetimepass==1.0.1
pypng==0.0.21
PyQRCode==1.2.1
Flask-HTTPAuth==4.1.0
pymongo==3.11.2